requirements.txt
flask
openai
python-docx
Pillow
gunicorn
flask-cors
python-dotenv
